requirements.txt
streamlit>=1.28.0
ccxt>=4.0.0
pandas>=2.0.0
plotly>=5.17.0
openpyxl>=3.1.2
APScheduler==3.10.4
psutil>=5.9.0
altair>=5.0.0
filelock>=3.12.0
requests>=2.31.0
